refactor: fix the failing test by moving QuickFilters types to types.tsx
